fix: rename input.svelte to camelCase annotation
