+++ src/module/combat/ResultsDialog.svelte
@@ -6,7 +6,7 @@
   import InputLabel from '@svelte/ui/inputLabel.svelte';
   import ModifiedAbilityInput from '@svelte/ui/modifiedAbilityInput.svelte';
   import { Attack } from './attack';
-  import Input from '@svelte/ui/Input.svelte';
+  import Input from '@svelte/ui/input.svelte';
 
   /**
    * @import { Defense } from './defense';
