+++ src/module/combat/defense/PhysicDefenseDialog.svelte
@@ -6,7 +6,7 @@
   import IconCheckBox from '@svelte/ui/iconCheckBox.svelte';
   import IconRange from '@svelte/ui/iconRange.svelte';
   import IconSwitch from '@svelte/ui/iconSwitch.svelte';
-  import Input from '@svelte/ui/Input.svelte';
+  import Input from '@svelte/ui/input.svelte';
   import InputLabel from '@svelte/ui/inputLabel.svelte';
   import ModifiedAbilityInput from '@svelte/ui/modifiedAbilityInput.svelte';
